Removed ProfileActivity and replacing with MainLanding Marked un-translatable strings proper back button for vaccine update activity
