+++ babyOne/app/src/main/java/com/example/babyone/VaccineUpdate.java
@@ -23,24 +23,7 @@ public class VaccineUpdate extends AppCompatActivity {
         btnback_2.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                // Create an instance of the profileFragment
-                profileFragment fragment = new profileFragment();
-
-                // Get the FragmentManager
-                FragmentManager fragmentManager = getSupportFragmentManager();
-
-                // Begin the fragment transaction
-                FragmentTransaction transaction = fragmentManager.beginTransaction();
-
-                // Replace the current fragment with the profileFragment
-                transaction.replace(R.id.profileFragment, fragment);
-
-                // Add the transaction to the back stack (optional)
-                transaction.addToBackStack(null);
-
-                // Commit the transaction
-                transaction.commit();
-
+                finish();
             }
         });
     }
