+++ babyOne/app/src/main/res/layout/activity_profile.xml
@@ -5,8 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:gravity="center"
-    tools:context=".ProfileActivity">
+    android:gravity="center" >
 
     <FrameLayout
         android:id="@+id/frame_layout"
