+++ babyOne/app/src/main/res/values-en/strings.xml
@@ -21,5 +21,4 @@
     <string name="strvalBTNBack">Back</string>
     <string name="strvalBTNNext">Next</string>
     <string name="strvalEBBTNSubmit">Submit</string>
-    <string name="strvalPlaceholder">placeholder</string>
 </resources>
