+++ babyOne/app/src/main/res/values-si/strings.xml
@@ -21,5 +21,4 @@
     <string name="strvalBTNBack">ආපසු</string>
     <string name="strvalBTNNext">ඉදිරියට</string>
     <string name="strvalEBBTNSubmit">තහවුරු කරන්න</string>
-    <string name="strvalPlaceholder">placeholder</string>
 </resources>
