+++ babyOne/app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
-    <string name="app_name">babyOne</string>
-    <string name="sign_up">Sign Up</string>
+    <string name="app_name" translatable="false">babyOne</string>
+    <string name="sign_up" translatable="false">Sign Up</string>
     <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="hello_blank_fragment" translatable="false">Hello blank fragment</string>
     <string name="strvalEBInfoBanner">Enter your baby\'s info</string>
     <string name="strvalEBInfoDetails">Please enter you baby\'s info. Only you and authorized personnel will be able to see this information.</string>
     <string name="strvalEBFullName">Full Name</string>
@@ -24,6 +24,6 @@
     <string name="strvalBTNBack">Back</string>
     <string name="strvalBTNNext">Next</string>
     <string name="strvalEBBTNSubmit">Submit</string>
-    <string name="strvalPlaceholder">placeholder</string>
-    <string name="strFASyringe"></string>
+    <string name="strvalPlaceholder" translatable="false">placeholder</string>
+    <string name="strFASyringe" translatable="false"></string>
 </resources>
